(feat) bundle cedict
