+++ .gitattributes
@@ -1,1 +1,2 @@
 web/lib/** linguist-vendored
+assets/** linguist-vendored
