+++ cnpy/api.py
@@ -8,10 +8,10 @@ from pprint import pprint
 import random
 from typing import Callable, Any, TypedDict, Optional
 
-from cnpy import quiz, cedict, tatoeba
+from cnpy import quiz, cedict, sentence
 from cnpy.db import db
 from cnpy.stats import make_stats
-from cnpy.dir import exe_root
+from cnpy.dir import assets_root, user_root
 
 
 srs = fsrs.FSRS()
@@ -26,7 +26,7 @@ class Api:
     web_ready: Callable
     web_window: Callable[[str, str, Optional[dict]], Any]
 
-    settings_path = exe_root / "user" / "settings.json"
+    settings_path = user_root / "settings.json"
     settings = UserSettings(levels=[])
 
     levels: dict[str, list[str]] = {}
@@ -37,7 +37,7 @@ class Api:
         if self.settings_path.exists():
             self.settings = json.loads(self.settings_path.read_text("utf-8"))
 
-        folder = exe_root / "assets" / "zhquiz-level"
+        folder = assets_root / "zhquiz-level"
         for f in folder.glob("**/*.txt"):
             self.levels[f.with_suffix("").name] = (
                 f.read_text(encoding="utf-8").rstrip().splitlines()
@@ -61,7 +61,7 @@ class Api:
     def start(self):
         quiz.load_db()
         cedict.load_db(self.web_log)
-        tatoeba.load_db(self.web_log)
+        sentence.load_db(self.web_log)
 
         self.web_ready()
 
@@ -147,7 +147,7 @@ class Api:
 
         vs = set()
 
-        path = exe_root / "user/vocab"
+        path = user_root / "vocab"
         path.mkdir(exist_ok=True)
 
         for f in path.glob("**/*.txt"):
@@ -184,7 +184,7 @@ class Api:
                 )
                 vs.add(v)
 
-        path = exe_root / "user/skip"
+        path = user_root / "skip"
         path.mkdir(exist_ok=True)
 
         for f in path.glob("**/*.txt"):
@@ -367,7 +367,7 @@ class Api:
         dict_entries.sort(key=sorter)
 
         sentences = [
-            tatoeba.load_db_entry(r)
+            sentence.load_db_entry(r)
             for r in db.execute(
                 """
             SELECT *
@@ -408,7 +408,7 @@ class Api:
                 ),
                 (v,),
             ):
-                r = tatoeba.load_db_entry(r)
+                r = sentence.load_db_entry(r)
                 if r["cmn"] not in prev_cmn:
                     sentences.append(r)
 
@@ -490,13 +490,13 @@ class Api:
         self.web_window(url, title, args)
 
     def load_file(self, f: str):
-        path = exe_root / "user" / f
+        path = user_root / f
         if path.exists():
-            return (exe_root / "user" / f).read_text(encoding="utf-8")
+            return (user_root / f).read_text(encoding="utf-8")
         return ""
 
     def save_file(self, f: str, txt: str):
-        (exe_root / "user" / f).write_text(txt, encoding="utf-8")
+        (user_root / f).write_text(txt, encoding="utf-8")
 
     def get_levels(self):
         return self.levels
